Refactor code in package imagebuildah
